+++ src/utility.h
@@ -24,8 +24,8 @@
 #define LD_UUID_SIZE 36
 
 bool LDi_sleepMilliseconds(const unsigned long milliseconds);
-bool LDi_getMonotonicMilliseconds(unsigned long *const resultMilliseconds);
-bool LDi_getUnixMilliseconds(unsigned long *const resultMilliseconds);
+bool LDi_getMonotonicMilliseconds(double *const resultMilliseconds);
+bool LDi_getUnixMilliseconds(double *const resultMilliseconds);
 bool LDi_randomHex(char *const buffer, const size_t bufferSize);
 bool LDi_UUIDv4(char *const buffer);
